combining PRs, removes credentials trait
