+++ src/LongRunning/Gapic/OperationsGapicClient.php
@@ -48,14 +48,13 @@ use Google\Cloud\Version;
 use Google\GAX\AgentHeaderDescriptor;
 use Google\GAX\ApiCallable;
 use Google\GAX\CallSettings;
-use Google\GAX\GrpcCredentialsHelper;
 use Google\GAX\PageStreamingDescriptor;
+use Google\GAX\GapicClientTrait;
 use Google\GAX\ValidationException;
 use Google\Longrunning\CancelOperationRequest;
 use Google\Longrunning\DeleteOperationRequest;
 use Google\Longrunning\GetOperationRequest;
 use Google\Longrunning\ListOperationsRequest;
-use Google\Longrunning\OperationsGrpcClient;
 
 /**
  * Service Description: Manages long-running operations with an API service.
@@ -95,6 +94,8 @@ use Google\Longrunning\OperationsGrpcClient;
  */
 class OperationsGapicClient
 {
+    use GapicClientTrait;
+
     /**
      * The default port of the service.
      */
@@ -113,7 +114,6 @@ class OperationsGapicClient
     private static $gapicVersion;
     private static $gapicVersionLoaded = false;
 
-    protected $grpcCredentialsHelper;
     protected $operationsStub;
     private $scopes;
     private $defaultCallSettings;
@@ -161,17 +161,8 @@ class OperationsGapicClient
      *
      *     @type string $serviceAddress Required. The domain name of the API remote host.
      *     @type mixed $port The port on which to connect to the remote host. Default 443.
-     *     @type \Grpc\Channel $channel
-     *           A `Channel` object to be used by gRPC. If not specified, a channel will be constructed.
-     *     @type \Grpc\ChannelCredentials $sslCreds
-     *           A `ChannelCredentials` object for use with an SSL-enabled channel.
-     *           Default: a credentials object returned from
-     *           \Grpc\ChannelCredentials::createSsl()
-     *           NOTE: if the $channel optional argument is specified, then this argument is unused.
-     *     @type bool $forceNewChannel
-     *           If true, this forces gRPC to create a new channel instead of using a persistent channel.
-     *           Defaults to false.
-     *           NOTE: if the $channel optional argument is specified, then this option is unused.
+     *     @type mixed $transport Optional, An instance of Google\Gax\TransportInterface or the
+     *           string "json" or "grpc". Determines the backend transport used to make the API call.
      *     @type \Google\Auth\CredentialsLoader $credentialsLoader
      *           A CredentialsLoader object created using the Google\Auth library.
      *     @type array $scopes Required. A string array of scopes to use when acquiring credentials.
@@ -209,6 +200,7 @@ class OperationsGapicClient
             'libName' => null,
             'libVersion' => null,
             'clientConfigPath' => __DIR__.'/../resources/operations_client_config.json',
+            'createOperationsStubFunction' => [$this, 'createStub'],
         ];
         $options = array_merge($defaultOptions, $options);
 
@@ -242,20 +234,7 @@ class OperationsGapicClient
                 );
 
         $this->scopes = $options['scopes'];
-
-        $createStubOptions = [];
-        if (array_key_exists('sslCreds', $options)) {
-            $createStubOptions['sslCreds'] = $options['sslCreds'];
-        }
-        $this->grpcCredentialsHelper = new GrpcCredentialsHelper($options);
-
-        $createOperationsStubFunction = function ($hostname, $opts, $channel) {
-            return new OperationsGrpcClient($hostname, $opts, $channel);
-        };
-        if (array_key_exists('createOperationsStubFunction', $options)) {
-            $createOperationsStubFunction = $options['createOperationsStubFunction'];
-        }
-        $this->operationsStub = $this->grpcCredentialsHelper->createStub($createOperationsStubFunction);
+        $this->operationsStub = call_user_func($options['createOperationsStubFunction'], $options);
     }
 
     /**
@@ -315,10 +294,7 @@ class OperationsGapicClient
             $this->descriptors['getOperation']
         );
 
-        return $callable(
-            $request,
-            [],
-            ['call_credentials_callback' => $this->createCredentialsCallback()]);
+        return $callable($request);
     }
 
     /**
@@ -410,10 +386,7 @@ class OperationsGapicClient
             $this->descriptors['listOperations']
         );
 
-        return $callable(
-            $request,
-            [],
-            ['call_credentials_callback' => $this->createCredentialsCallback()]);
+        return $callable($request);
     }
 
     /**
@@ -478,10 +451,7 @@ class OperationsGapicClient
             $this->descriptors['cancelOperation']
         );
 
-        return $callable(
-            $request,
-            [],
-            ['call_credentials_callback' => $this->createCredentialsCallback()]);
+        return $callable($request);
     }
 
     /**
@@ -540,10 +510,7 @@ class OperationsGapicClient
             $this->descriptors['deleteOperation']
         );
 
-        return $callable(
-            $request,
-            [],
-            ['call_credentials_callback' => $this->createCredentialsCallback()]);
+        return $callable($request);
     }
 
     /**
@@ -555,10 +522,5 @@ class OperationsGapicClient
     public function close()
     {
         $this->operationsStub->close();
-    }
-
-    private function createCredentialsCallback()
-    {
-        return $this->grpcCredentialsHelper->createCallCredentialsCallback();
     }
 }
