+++ src/LongRunning/OperationsTransportInterface.php
@@ -48,7 +48,6 @@ use Google\GAX\AgentHeaderDescriptor;
 use Google\GAX\ApiCallable;
 use Google\GAX\CallSettings;
 use Google\GAX\GrpcConstants;
-use Google\GAX\GrpcCredentialsHelper;
 use Google\GAX\PageStreamingDescriptor;
 use Google\GAX\ValidationException;
 use Google\Longrunning\CancelOperationRequest;
